+++ boot.dev/aggrss/sql/queries/feed_follows.sql
@@ -4,7 +4,7 @@ VALUES ($1, $2, $3, $4, $5)
 RETURNING *;
 
 -- name: DeleteFeedFollowByID :exec
-DELETE FROM feed_follows WHERE id = $1;
+DELETE FROM feed_follows WHERE id = $1 AND user_id = $2;
 
 -- name: ListFeedFollowsByUser :many
 SELECT *
